// src/api/generated/model/consumableDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * GearFlow.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ConsumableDtoUnitPriceCents } from './consumableDtoUnitPriceCents';
import type { ConsumableDtoQuantity } from './consumableDtoQuantity';
import type { ConsumableDtoReservedQuantity } from './consumableDtoReservedQuantity';

export interface ConsumableDto {
  id: string;
  name: string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  unitPriceCents: ConsumableDtoUnitPriceCents;
  /** @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$ */
  quantity: ConsumableDtoQuantity;
  /** @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$ */
  reservedQuantity: ConsumableDtoReservedQuantity;
  belowMinimum: boolean;
}