// src/api/generated/model/partDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * GearFlow.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { PartDtoPriceCents } from './partDtoPriceCents';
import type { PartDtoQuantity } from './partDtoQuantity';
import type { PartDtoReservedQuantity } from './partDtoReservedQuantity';

export interface PartDto {
  id: string;
  name: string;
  description: string;
  partNumber: string;
  manufacturer: string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  priceCents: PartDtoPriceCents;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  quantity: PartDtoQuantity;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  reservedQuantity: PartDtoReservedQuantity;
  belowMinimum: boolean;
}